Added NavigationController.
Pulled everything out of UAAppDelegate and put it in a new UANavigationController class. UANC has a container view into which UATV views are added as and when they are needed.
Added a header view to xib used by UATV, which contains a back button and a title text field, exposed through properties to be set by the navigation controller.
Table view now pushed on double click rather than single click.
When we push a table view, we now position it relative to the visible view controller's view and clip the rest thanks to the contained view of UANC.