Update animations.
Now animate from half-way into frame. What actually happens on iOS is that the tableview itself animates the full width of the frame (i.e. fully offscreen to fully onscreen). The navigation bar of the current tableview animates and fades out, while the navigation bar for the incoming tableview fades in in place. Replicating this would require a reworking of the structure of the tableview/nib. A project for later on.